Added the folder I have been working on to implement imaging, starting with basic raster scanning. Of note is the fpga labview file explore-z-axis.vi. This vi implements a state machine with the following structure
1). Assume tip is above the sample surface, and just hold it there. Boolean trigger --> state 2
2). Lower tip. The error signal remains relatively constant until we engage the sample. Thus, as long
	as the error signal is within some bounds (defined be TOL and the initial error signal), we keep lowing.
	Once error signal leaves the bounds, we trigger a boolean and transition --> state 3.
3). PID. Given a reference setpoint, implements PI control on z-axis to regulate about setpoint. Triggering in
	this state should return us to state (1).

I have also used this same state machine architecture to initialize a vi for doing raster scanning, called raster-scanning.vi.
That vi is completely untested though. I'm not sure if processing the data is better done on the FPGA or offline/on the host.
Right now it is setup to feed in a FIFO signal for x-y ramps/triangle waves and just export the x-y-z data back to the host
at every sample instant via another FIFO buffer.