Fixing bugs to try to get the multiple scans to work
-- in the state machine, when the scan ends, (state 5) need to tell the
FIFO to not acquire new data.
-- Added accumulator resets to the PI controllers, trying to track down
this strange bug where things appear unstable (in the sense that the
control exceeds my bound).
-- It appears that problem was due to the return_axis code NOT getting
executed. I think my "trick" of using a what I called a phony_enable
didnt work, and labview decided there was no data dependency and so
skipped straight to the shutdown.

-- Need to close file in process_scan_data